requirements.txt
mysql-connector-python==9.6.0
psycopg2==2.9.11
psycopg2-binary==2.9.11
python-dotenv==1.2.1
